requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
djangorestframework==3.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
